rolemembershipcache: extract cache to new package
The caching logic was getting to be quite large, and it already was
nicely contained in an easy to move struct.

This is a purely mechanical refactor.

Release note: None
