+++ pkg/BUILD.bazel
@@ -2103,6 +2103,7 @@ GO_TARGETS = [
     "//pkg/sql/regionliveness:regionliveness",
     "//pkg/sql/regions:regions",
     "//pkg/sql/regions:regions_test",
+    "//pkg/sql/rolemembershipcache:rolemembershipcache",
     "//pkg/sql/roleoption:roleoption",
     "//pkg/sql/row:row",
     "//pkg/sql/row:row_test",
