+++ pkg/server/BUILD.bazel
@@ -246,6 +246,7 @@ go_library(
         "//pkg/sql/privilege",
         "//pkg/sql/querycache",
         "//pkg/sql/rangeprober",
+        "//pkg/sql/rolemembershipcache",
         "//pkg/sql/roleoption",
         "//pkg/sql/scheduledlogging",
         "//pkg/sql/schemachanger/scdeps",
