+++ pkg/sql/BUILD.bazel
@@ -435,7 +435,6 @@ go_library(
         "//pkg/sql/isql",
         "//pkg/sql/lex",
         "//pkg/sql/lexbase",
-        "//pkg/sql/memsize",
         "//pkg/sql/mutations",
         "//pkg/sql/oidext",
         "//pkg/sql/opt",
@@ -469,6 +468,7 @@ go_library(
         "//pkg/sql/querycache",
         "//pkg/sql/regionliveness",
         "//pkg/sql/regions",
+        "//pkg/sql/rolemembershipcache",
         "//pkg/sql/roleoption",
         "//pkg/sql/row",
         "//pkg/sql/rowcontainer",
@@ -587,7 +587,6 @@ go_library(
         "//pkg/util/startup",
         "//pkg/util/stop",
         "//pkg/util/syncutil",
-        "//pkg/util/syncutil/singleflight",
         "//pkg/util/timeutil",
         "//pkg/util/timeutil/pgdate",
         "//pkg/util/tochar",
