+++ pkg/sql/rolemembershipcache/BUILD.bazel
@@ -1,0 +1,28 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "rolemembershipcache",
+    srcs = ["cache.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/sql/rolemembershipcache",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/keys",
+        "//pkg/roachpb",
+        "//pkg/security/username",
+        "//pkg/sql/catalog/descpb",
+        "//pkg/sql/catalog/descs",
+        "//pkg/sql/catalog/lease",
+        "//pkg/sql/isql",
+        "//pkg/sql/memsize",
+        "//pkg/sql/sem/tree",
+        "//pkg/sql/sessiondata",
+        "//pkg/sql/sqlerrors",
+        "//pkg/util/hlc",
+        "//pkg/util/log",
+        "//pkg/util/mon",
+        "//pkg/util/stop",
+        "//pkg/util/syncutil",
+        "//pkg/util/syncutil/singleflight",
+        "@com_github_cockroachdb_errors//:errors",
+    ],
+)
